Gets, Create and Custom Exceptions

+++ ProjectSimple.Application/Interfaces/IGenericRepository.cs
@@ -7,6 +7,6 @@ public interface IGenericRepository<T> where T : BaseEntity
     Task<T> CreateAsync(T entity);
     Task<T> UpdateAsync(T entity);
     Task<T> DeleteAsync(T entity);
-    Task<T> GetAsync(int Id);
+    Task<T> GetAsync(long Id);
     Task<IReadOnlyList<T>> GetAllAsync();
 }
